More fixes to flatfile serializer.

+++ OsmSharp.Routing/Graph/Serialization/FlatfileSerializer.cs
@@ -95,21 +95,21 @@ namespace OsmSharp.Routing.Graph.Serialization
             var sizeBytes = new byte[8];
             stream.Read(sizeBytes, 0, 8);
             var position = stream.Position;
-            var size = BitConverter.ToInt32(sizeBytes, 0);
+            var size = BitConverter.ToInt64(sizeBytes, 0);
             this.DeserializeVertices(stream, size, graph);
             stream.Seek(position + size, System.IO.SeekOrigin.Begin);
 
             // deserialize edges.
             stream.Read(sizeBytes, 0, 8);
             position = stream.Position;
-            size = BitConverter.ToInt32(sizeBytes, 0);
+            size = BitConverter.ToInt64(sizeBytes, 0);
             this.DeserializeEdges(stream, size, graph);
             stream.Seek(position + size, System.IO.SeekOrigin.Begin);
 
             // deserialize tags.
             stream.Read(sizeBytes, 0, 8);
             position = stream.Position;
-            size = BitConverter.ToInt32(sizeBytes, 0);
+            size = BitConverter.ToInt64(sizeBytes, 0);
             this.DeserializeTags(stream, size, tagsCollectionIndex);
             stream.Seek(position + size, System.IO.SeekOrigin.Begin);
